code quality: try-catch, error boundaries, DRY, SOLID form splits, schema extraction
- Add try-catch to deleteInvoice, markAsPaid, toggleRecurringInvoice,
  deleteRecurringInvoice, /api/reports/export, and /api/dashboard/chart-data
- Add error.tsx to all missing dashboard sub-routes: invoices, clients,
  billing, profile, recurring-invoices, reports, invoices/[invoiceId],
  clients/[clientId]
- Extract sendInvoiceEmail() helper in invoices.ts to eliminate 95-line
  duplication between createInvoice and editInvoice
- Replace Intl.DateTimeFormat with formatDate.long() in recurringInvoices.ts
- Split ClientForm (699 lines) into BasicInfoTab, AddressesTab,
  ContactPersonsTab, CustomFieldsTab sub-components
- Split ProfileForm (409 lines) into PersonalTab, CompanyTab, BankTab
- Split lib/zodSchemas.ts into lib/schemas/{invoice,client,onboarding,recurringInvoice}.ts
  with re-exports from zodSchemas.ts for backward compat
- Remove userData prop drilling from clients DataTable/columns now that
  InvoiceDialog reads user data from UserContext
- Fix usage.test.ts TypeScript cast via unknown intermediate
