+++ components/client-form/clientForm.tsx
@@ -4,34 +4,20 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
 import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
+import { Form } from "@/components/ui/form";
 import { toast } from "sonner";
 import { Client } from "@/app/dashboard/clients/columns";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Card, CardContent } from "@/components/ui/card";
-import { Plus, Trash2 } from "lucide-react";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { clientFormSchema } from "@/lib/zodSchemas";
 import { createClient, editClient } from "@/app/actions/clients";
 import { useRouter } from "next/navigation";
 import SubmitButton from "@/components/submit-button/SubmitButton";
 import { ConfirmDialog } from "@/components/ui/confirm-dialog";
 import { useState } from "react";
+import { BasicInfoTab } from "./BasicInfoTab";
+import { AddressesTab } from "./AddressesTab";
+import { ContactPersonsTab } from "./ContactPersonsTab";
+import { CustomFieldsTab } from "./CustomFieldsTab";
 
 type ClientFormValues = z.infer<typeof clientFormSchema>;
 
@@ -126,10 +112,8 @@ export function ClientForm({ client, onClose, onSuccess }: ClientFormProps) {
       setIsLoading(true);
       const formData = new FormData();
 
-      // Add all form fields to FormData
       Object.entries(data).forEach(([key, value]) => {
         if (Array.isArray(value)) {
-          // For arrays, append each item with an index
           formData.append(key, JSON.stringify(value));
         } else if (value !== null && value !== undefined) {
           formData.append(key, value.toString());
@@ -142,9 +126,7 @@ export function ClientForm({ client, onClose, onSuccess }: ClientFormProps) {
 
       if (result.status === "error") {
         console.error("Form validation errors:", result.error);
-        toast.error(
-          "Failed to update client. Please check the form for errors."
-        );
+        toast.error("Failed to update client. Please check the form for errors.");
         return;
       }
 
@@ -154,9 +136,7 @@ export function ClientForm({ client, onClose, onSuccess }: ClientFormProps) {
       onClose?.();
     } catch (error) {
       console.error("Submission error:", error);
-      toast.error(
-        error instanceof Error ? error.message : "Something went wrong"
-      );
+      toast.error(error instanceof Error ? error.message : "Something went wrong");
     } finally {
       setIsLoading(false);
     }
@@ -174,495 +154,19 @@ export function ClientForm({ client, onClose, onSuccess }: ClientFormProps) {
           </TabsList>
 
           <TabsContent value="basic" className="space-y-4">
-            <Card>
-              <CardContent className="pt-6">
-                <div className="grid gap-4 md:grid-cols-2">
-                  <FormField
-                    control={form.control}
-                    name="name"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Name</FormLabel>
-                        <FormControl>
-                          <Input placeholder="Client name" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="email"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Email</FormLabel>
-                        <FormControl>
-                          <Input
-                            type="email"
-                            placeholder="client@example.com"
-                            {...field}
-                            value={field.value || ""}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="phone"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Phone</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="Phone number"
-                            {...field}
-                            value={field.value || ""}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="taxId"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Tax ID</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="Tax ID"
-                            {...field}
-                            value={field.value || ""}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="website"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Website</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="https://example.com"
-                            {...field}
-                            value={field.value || ""}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="category"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Category</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="Client category"
-                            {...field}
-                            value={field.value || ""}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <div className="col-span-1 sm:col-span-2">
-                    <FormField
-                      control={form.control}
-                      name="notes"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Notes</FormLabel>
-                          <FormControl>
-                            <Textarea
-                              placeholder="Additional notes about the client"
-                              {...field}
-                              value={field.value || ""}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-                </div>
-              </CardContent>
-            </Card>
+            <BasicInfoTab form={form} />
           </TabsContent>
 
           <TabsContent value="addresses" className="space-y-4">
-            <Card>
-              <CardContent className="pt-6">
-                {form.watch("addresses")?.map((_, index) => (
-                  <div
-                    key={index}
-                    className="grid gap-4 md:grid-cols-2 mb-8 pb-8 border-b last:mb-0 last:pb-0 last:border-0"
-                  >
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.type`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Address Type</FormLabel>
-                          <Select
-                            onValueChange={field.onChange}
-                            defaultValue={field.value}
-                          >
-                            <FormControl>
-                              <SelectTrigger>
-                                <SelectValue placeholder="Select address type" />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent>
-                              <SelectItem value="BILLING">Billing</SelectItem>
-                              <SelectItem value="SHIPPING">Shipping</SelectItem>
-                              <SelectItem value="OTHER">Other</SelectItem>
-                            </SelectContent>
-                          </Select>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.street`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Street</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Street address" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.city`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>City</FormLabel>
-                          <FormControl>
-                            <Input placeholder="City" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.state`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>State/Province</FormLabel>
-                          <FormControl>
-                            <Input placeholder="State/Province" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.country`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Country</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Country" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`addresses.${index}.zipCode`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>ZIP/Postal Code</FormLabel>
-                          <FormControl>
-                            <Input placeholder="ZIP/Postal code" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <div className="col-span-1 sm:col-span-2 flex justify-end gap-2">
-                      {index > 0 && (
-                        <Button
-                          type="button"
-                          variant="destructive"
-                          size="sm"
-                          onClick={() => {
-                            const addresses = form.getValues("addresses");
-                            form.setValue(
-                              "addresses",
-                              addresses.filter((_, i) => i !== index)
-                            );
-                          }}
-                        >
-                          <Trash2 className="h-4 w-4 mr-2" />
-                          Remove Address
-                        </Button>
-                      )}
-                    </div>
-                  </div>
-                ))}
-
-                <Button
-                  type="button"
-                  variant="outline"
-                  className="mt-4"
-                  onClick={() => {
-                    const addresses = form.getValues("addresses");
-                    form.setValue("addresses", [
-                      ...addresses,
-                      {
-                        type: "BILLING",
-                        street: "",
-                        city: "",
-                        state: "",
-                        country: "",
-                        zipCode: "",
-                        isDefault: false,
-                      },
-                    ]);
-                  }}
-                >
-                  <Plus className="h-4 w-4 mr-2" />
-                  Add Address
-                </Button>
-              </CardContent>
-            </Card>
+            <AddressesTab form={form} />
           </TabsContent>
 
           <TabsContent value="contacts" className="space-y-4">
-            <Card>
-              <CardContent className="pt-6">
-                {form.watch("contactPersons")?.map((_, index) => (
-                  <div
-                    key={index}
-                    className="grid gap-4 md:grid-cols-2 mb-8 pb-8 border-b last:mb-0 last:pb-0 last:border-0"
-                  >
-                    <FormField
-                      control={form.control}
-                      name={`contactPersons.${index}.firstName`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>First Name</FormLabel>
-                          <FormControl>
-                            <Input placeholder="First name" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`contactPersons.${index}.lastName`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Last Name</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Last name" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`contactPersons.${index}.email`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Email</FormLabel>
-                          <FormControl>
-                            <Input
-                              type="email"
-                              placeholder="contact@example.com"
-                              {...field}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`contactPersons.${index}.phone`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Phone</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Phone number" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`contactPersons.${index}.position`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Position</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Job position" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <div className="col-span-1 sm:col-span-2 flex justify-end">
-                      <Button
-                        type="button"
-                        variant="destructive"
-                        size="sm"
-                        onClick={() => {
-                          const contacts = form.getValues("contactPersons");
-                          form.setValue(
-                            "contactPersons",
-                            contacts.filter((_, i) => i !== index)
-                          );
-                        }}
-                      >
-                        <Trash2 className="h-4 w-4 mr-2" />
-                        Remove Contact
-                      </Button>
-                    </div>
-                  </div>
-                ))}
-
-                <Button
-                  type="button"
-                  variant="outline"
-                  className="mt-4"
-                  onClick={() => {
-                    const contacts = form.getValues("contactPersons");
-                    form.setValue("contactPersons", [
-                      ...contacts,
-                      {
-                        firstName: "",
-                        lastName: "",
-                        email: "",
-                        phone: "",
-                        position: "",
-                        isPrimary: !contacts.length,
-                      },
-                    ]);
-                  }}
-                >
-                  <Plus className="h-4 w-4 mr-2" />
-                  Add Contact Person
-                </Button>
-              </CardContent>
-            </Card>
+            <ContactPersonsTab form={form} />
           </TabsContent>
 
           <TabsContent value="custom" className="space-y-4">
-            <Card>
-              <CardContent className="pt-6">
-                {form.watch("customFields")?.map((_, index) => (
-                  <div
-                    key={index}
-                    className="grid gap-4 md:grid-cols-2 mb-8 pb-8 border-b last:mb-0 last:pb-0 last:border-0"
-                  >
-                    <FormField
-                      control={form.control}
-                      name={`customFields.${index}.key`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Field Name</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Field name" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name={`customFields.${index}.value`}
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Field Value</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Field value" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <div className="col-span-1 sm:col-span-2 flex justify-end">
-                      <Button
-                        type="button"
-                        variant="destructive"
-                        size="sm"
-                        onClick={() => {
-                          const fields = form.getValues("customFields");
-                          form.setValue(
-                            "customFields",
-                            fields.filter((_, i) => i !== index)
-                          );
-                        }}
-                      >
-                        <Trash2 className="h-4 w-4 mr-2" />
-                        Remove Field
-                      </Button>
-                    </div>
-                  </div>
-                ))}
-
-                <Button
-                  type="button"
-                  variant="outline"
-                  className="mt-4"
-                  onClick={() => {
-                    const fields = form.getValues("customFields");
-                    form.setValue("customFields", [
-                      ...fields,
-                      {
-                        key: "",
-                        value: "",
-                      },
-                    ]);
-                  }}
-                >
-                  <Plus className="h-4 w-4 mr-2" />
-                  Add Custom Field
-                </Button>
-              </CardContent>
-            </Card>
+            <CustomFieldsTab form={form} />
           </TabsContent>
         </Tabs>
