+++ components/profile-form/BankTab.tsx
@@ -1,0 +1,114 @@
+"use client";
+
+import { UseFormReturn } from "react-hook-form";
+import * as z from "zod";
+import { Card, CardContent } from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { onboardingSchema } from "@/lib/zodSchemas";
+
+type ProfileFormValues = z.infer<typeof onboardingSchema>;
+
+interface BankTabProps {
+  form: UseFormReturn<ProfileFormValues>;
+}
+
+export function BankTab({ form }: BankTabProps) {
+  return (
+    <Card>
+      <CardContent className="pt-6">
+        <div className="grid gap-4">
+          <FormField
+            control={form.control}
+            name="bankName"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Bank Name</FormLabel>
+                <FormControl>
+                  <Input placeholder="Bank Name" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="bankAccountName"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Account Name</FormLabel>
+                <FormControl>
+                  <Input placeholder="Account Name" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="bankAccountNumber"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Account Number</FormLabel>
+                <FormControl>
+                  <Input placeholder="Account Number" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="bankSwiftCode"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Swift Code</FormLabel>
+                <FormControl>
+                  <Input placeholder="Swift Code" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="bankIBAN"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>IBAN</FormLabel>
+                <FormControl>
+                  <Input placeholder="IBAN" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="bankAddress"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Bank Address</FormLabel>
+                <FormControl>
+                  <Input placeholder="Bank Address" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+      </CardContent>
+    </Card>
+  );
+}
