+++ components/profile-form/ProfileForm.tsx
@@ -13,19 +13,14 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
+import { Form } from "@/components/ui/form";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import SubmitButton from "@/components/submit-button/SubmitButton";
 import { updateProfile } from "@/app/actions/profile";
 import { onboardingSchema } from "@/lib/zodSchemas";
+import { PersonalTab } from "./PersonalTab";
+import { CompanyTab } from "./CompanyTab";
+import { BankTab } from "./BankTab";
 
 type ProfileFormValues = z.infer<typeof onboardingSchema>;
 
@@ -94,7 +89,6 @@ export function ProfileForm({
       setIsLoading(true);
       const formData = new FormData();
 
-      // Add all form fields to FormData
       Object.entries(data).forEach(([key, value]) => {
         if (value !== null && value !== undefined) {
           formData.append(key, value.toString());
@@ -121,16 +115,10 @@ export function ProfileForm({
       <CardHeader className="flex flex-row items-center justify-between">
         <div>
           <CardTitle>Profile Settings</CardTitle>
-          <CardDescription>
-            Update your personal information here
-          </CardDescription>
+          <CardDescription>Update your personal information here</CardDescription>
         </div>
         <div className="flex-shrink-0 w-fit">
-          <SubmitButton
-            text="Update Profile"
-            isLoading={isLoading}
-            form="profile-form"
-          />
+          <SubmitButton text="Update Profile" isLoading={isLoading} form="profile-form" />
         </div>
       </CardHeader>
       <CardContent>
@@ -148,257 +136,15 @@ export function ProfileForm({
               </TabsList>
 
               <TabsContent value="basic" className="space-y-4">
-                <Card>
-                  <CardContent className="pt-6">
-                    <div className="grid md:grid-cols-2 gap-4">
-                      <FormField
-                        control={form.control}
-                        name="firstName"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>First Name</FormLabel>
-                            <FormControl>
-                              <Input placeholder="John" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="lastName"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Last Name</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Doe" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="grid gap-4 mt-4">
-                      <div className="grid gap-2">
-                        <FormLabel>Email</FormLabel>
-                        <Input value={email} disabled className="bg-muted" />
-                      </div>
-
-                      <FormField
-                        control={form.control}
-                        name="address"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Address</FormLabel>
-                            <FormControl>
-                              <Input placeholder="123 Street Name" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </CardContent>
-                </Card>
+                <PersonalTab form={form} email={email} />
               </TabsContent>
 
               <TabsContent value="company" className="space-y-4">
-                <Card>
-                  <CardContent className="pt-6">
-                    <div className="grid gap-4">
-                      <FormField
-                        control={form.control}
-                        name="companyName"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Company Name</FormLabel>
-                            <FormControl>
-                              <Input
-                                placeholder="Your Company Name"
-                                {...field}
-                              />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="companyEmail"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Company Email</FormLabel>
-                            <FormControl>
-                              <Input
-                                placeholder="company@example.com"
-                                type="email"
-                                {...field}
-                              />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="companyAddress"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Company Address</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Company Address" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="companyTaxId"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Company Tax ID</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Tax ID Number" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="companyLogoUrl"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Company Logo URL</FormLabel>
-                            <FormControl>
-                              <Input
-                                placeholder="https://example.com/logo.png"
-                                {...field}
-                              />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="stampsUrl"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Stamps URL</FormLabel>
-                            <FormControl>
-                              <Input
-                                placeholder="https://example.com/stamps.png"
-                                {...field}
-                              />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </CardContent>
-                </Card>
+                <CompanyTab form={form} />
               </TabsContent>
 
               <TabsContent value="bank" className="space-y-4">
-                <Card>
-                  <CardContent className="pt-6">
-                    <div className="grid gap-4">
-                      <FormField
-                        control={form.control}
-                        name="bankName"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Bank Name</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Bank Name" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="bankAccountName"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Account Name</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Account Name" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="bankAccountNumber"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Account Number</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Account Number" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="bankSwiftCode"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Swift Code</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Swift Code" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="bankIBAN"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>IBAN</FormLabel>
-                            <FormControl>
-                              <Input placeholder="IBAN" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="bankAddress"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Bank Address</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Bank Address" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </CardContent>
-                </Card>
+                <BankTab form={form} />
               </TabsContent>
             </Tabs>
           </form>
